git: fmt-merge-msg
